deps: upgrade+update
`yarn audit` reports 16k issues with our packages. This commit brings it
down to zero. We should keep running `yarn upgrade` in the future, in
order to develop against the same dependencies as our clients.

Since `sinon` was bumped, some old behaviour seems to have changed.
Fixed the parts of the tests where the failures occured.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>